deduplicate functions in nodejs scripts
